requirements.txt
pytest # ~=7.2.1
numpy~=1.24.4
pandas~=2.0.3
python-dotenv~=1.0.0
dask~=2023.4.1
sqlalchemy~=2.0.23
psycopg2~=2.9.3

seaborn~=0.13.0
matplotlib # ~=3.5.3
cursor~=1.3.5
pyodbc~=5.0.1
